moving RpRegistrationFilter to ms-rest
