+++ runtime/ms-rest-azure/package.json
@@ -5,7 +5,7 @@
     "email": "azsdkteam@microsoft.com",
     "url": "https://github.com/Azure/azure-sdk-for-node"
   },
-  "version": "2.5.3",
+  "version": "2.5.4",
   "description": "Client Runtime for Node.js Azure client libraries generated using AutoRest",
   "tags": [
     "node",
@@ -25,13 +25,13 @@
   "types": "./index.d.ts",
   "license": "MIT",
   "dependencies": {
-    "async": "2.5.0",
-    "uuid": "^3.1.0",
+    "async": "2.6.0",
+    "uuid": "^3.2.1",
     "adal-node": "^0.1.27",
-    "ms-rest": "^2.3.0",
-    "moment": "~2.18.1",
-    "@types/uuid": "^3.4.2",
-    "@types/node": "^8.0.25"
+    "ms-rest": "^2.3.1",
+    "moment": "^2.20.1",
+    "@types/uuid": "^3.4.3",
+    "@types/node": "^9.4.6"
   },
   "devDependencies": {
     "@types/mocha": "^2.2.40",
@@ -41,7 +41,7 @@
     "should": "5.2.0",
     "sinon": "^2.3.8",
     "nock": "^9.0.14",
-    "typescript": "^2.6.1"
+    "typescript": "^2.7.2"
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-node/runtime/ms-rest-azure",
   "repository": {
