+++ runtime/ms-rest-azure/test/testlist.txt
@@ -1,6 +1,5 @@
 azureServiceClientTests.js
 azureEnvironmentTests.js
-rpRegistrationFilterTests.js
 credTests.js
 msiVmTokenCredentialTests.js
 msiAppServiceTokenCredentialTests.js
