+++ runtime/ms-rest/test/rpRegistrationFilterTests.js
